requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
"packaging==21.3
pip==22.1.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
requests==2.28.1
selenium-page-factory==2.5
selenium==4.7.0
setuptools==60.2.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio-websocket==0.9.2
trio==0.21.0
urllib3==1.26.10
webdriver-manager==3.8.1
wheel==0.37.1
wsproto==1.1.0
undetected-chromedriver-3.2.1
websockets-10.4
